Continued work on tests

+++ bin/setup_hooks.sh
@@ -2,4 +2,4 @@
 
 pip3 install flake8
 cd ./.git/hooks || exit 1
-ln -s ../../bin/pre-commit pre-commit
+[[ ! -L ./pre-commit  ]] && ln -s ../../bin/pre-commit pre-commit || exit 0
